Move rancher types under separate external API package
