+++ api/rancher/management/v3/clusterregistrationtoken.go
