+++ internal/controllers/patch_kcfg_controller.go
@@ -35,7 +35,7 @@ import (
 	"sigs.k8s.io/cluster-api/controllers/external"
 	"sigs.k8s.io/cluster-api/util/predicates"
 
-	provisioningv1 "github.com/rancher/turtles/internal/rancher/provisioning/v1"
+	provisioningv1 "github.com/rancher/turtles/api/rancher/provisioning/v1"
 	turtlespredicates "github.com/rancher/turtles/util/predicates"
 )
